LayoutTests/platform/win/fast/canvas/canvas-fillRect-gradient-shadow-expected.txt

Ensure correct behavior of canvas with fillRect using a gradient fillStyle and a shadow

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 255. Was 0.
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 76 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 76 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 76 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] is not around 76 (actual: 0)
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 25
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 10
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
FAIL d[0] should be 255. Was 0.
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is smaller than 15
PASS successfullyParsed is true

TEST COMPLETE
